[PS] Clean components, files
