+++ payment-service/src/services/webhookService.ts
@@ -1,6 +1,7 @@
 import { PrismaClient, PaymentStatus, TransactionType } from '@prisma/client';
 import { Logger } from '../utils/logger';
-import { OmiseWebhookEvent } from '../types/payment';
+import { OmiseWebhookEvent, WebhookResponse } from '../types/payment';
+import { Request, Response } from 'express';
 import { v4 as uuidv4 } from 'uuid';
 
 const prisma = new PrismaClient();
@@ -14,6 +15,77 @@ export interface WebhookProcessingResult {
 }
 
 export class WebhookService {
+  /**
+   * Express handler wrapper to process incoming Omise webhook HTTP requests.
+   * Keeps HTTP / transport concerns out of the server bootstrap file.
+   */
+  async handleOmiseWebhookHttp(req: Request, res: Response): Promise<void> {
+    try {
+      const webhookEvent: OmiseWebhookEvent = req.body;
+
+      Logger.success('Omise webhook received', {
+        event_id: webhookEvent.id,
+        event_type: webhookEvent.key,
+        livemode: webhookEvent.livemode,
+        data_object: webhookEvent.data?.object,
+        data_id: webhookEvent.data?.id,
+        amount: webhookEvent.data?.amount,
+        currency: webhookEvent.data?.currency,
+        status: webhookEvent.data?.status,
+        created_at: webhookEvent.created_at
+      });
+
+      const result = await this.processWebhookEvent(webhookEvent);
+
+      if (result.success) {
+        const response: WebhookResponse = {
+          received: true,
+          timestamp: new Date().toISOString(),
+          event_type: webhookEvent.key || 'unknown',
+          event_id: webhookEvent.id || 'unknown',
+          message: result.message
+        };
+
+        Logger.success('Webhook processed successfully', {
+          event_id: webhookEvent.id,
+          payment_id: result.paymentId,
+          updated_status: result.updatedStatus,
+          message: result.message
+        });
+
+        res.status(200).json(response);
+      } else {
+        Logger.error('Webhook processing failed', {
+          event_id: webhookEvent.id,
+          error: result.error,
+          message: result.message
+        });
+
+        const errorResponse: WebhookResponse = {
+          received: false,
+          timestamp: new Date().toISOString(),
+          event_type: webhookEvent.key || 'unknown',
+          event_id: webhookEvent.id || 'unknown',
+          message: result.message || 'Failed to process webhook'
+        };
+
+        res.status(400).json(errorResponse);
+      }
+    } catch (error) {
+      Logger.error('Error processing Omise webhook', error);
+
+      const errorResponse: WebhookResponse = {
+        received: false,
+        timestamp: new Date().toISOString(),
+        event_type: 'error',
+        event_id: 'unknown',
+        message: 'Error processing webhook'
+      };
+
+      res.status(500).json(errorResponse);
+    }
+  }
+
   /**
    * Process incoming Omise webhook events
    */
